+++ crates/ronami-core/src/types/chat_administrator_rights.rs
@@ -64,4 +64,8 @@ pub struct ChatAdministratorRights {
     /// `true`, if the administrator can manage direct messages of the channel
     /// and decline suggested posts; for channels only
     pub can_manage_direct_messages: bool,
+
+    /// `true`, if the administrator can edit the tags of regular members; for
+    /// groups and supergroups only
+    pub can_manage_tags: Option<bool>,
 }